Fix CODE-HEADER-FROM-PC even more
A previous fix for crashing in backtraces merely disabled GC in the heap scan,
which is not only not great, but also not adequate.

Due to reuse of holes in immobile space, the sizetab[] functions might cause
reading of random words as object headers while the allocation is in progress.
Incidentally - and not as if it matters to this change - by why are tests in
compiler.pure performing backtrace so much? It seems weird, though "helpful" in
exposing the bug.

The situation is better in dynamic space- if another thread allocates code,
though there is a small chance of a glitch when the first word of the new
object is written and not the second- the heap-walking thread will compute that
size as being too small, followed by zeros which look like cons cells.