Add support to 'run-tests' for compiling in multiple threads
This changes only the test driver, without removing *world-lock*.
The environment variable SBCL_TEST_PARALLEL controls the number of
threads to create.
Produce a "test.log" file containing the elapsed time and namestring
of each '.pure' file run. (impure and/or cload files are not counted yet)

Some tests which use ASSERT-NO-CONSING are manually annotated to run
serially since presence of other threads would contaminate the test.
(The mechanism for asserting no-consing should be based on a compile-time
event recorder of whether we called the ALLOCATION macro, so that we don't
have to invoke the the generated code to decide if any allocation code was
produced. But this is a separate concern)