Futz with the first two words in code objects
With this change, the code header word contains the complete object size.
The boxed portion size (measured in bytes) is in the following word.

As evident from the number of deleted lines, it is easier to do pointer
arithmetic this way because all of the COMPUTE-FUN vops want bytes, not words.
For cheneygc, all of the backend-specific allocate-code-object vops are gone,
and the allocator essentially reduces to VAR-ALLOC.
For gencgc the change is similar but we still go through C code for reasons.

In either case, object sizing is trivial, involving no adding or rounding.