Remove an argument from alloc-immobile-fixedobj
The 2 'size' arguments were for setting a code object's size, but the boxed
size can safely be assigned after the pseudoatomic allocation sequence.
Given the newfound freedom to reassign the boxed size (subject to not making
it smaller than the number of pointer words) GC need not scavenge the 'fixups'
slot of an immobile trampoline. The slot is logically in the unboxed bytes.