Remove !DEFVAR and !DEFPARAMETER cold-init helpers
Occam's razor at work - during cold-load they both reduce to the same thing.
The distinction is largely meaningless because logically the symbol in DEFVAR
can't have a prior value - there is no prior state. Thus the macros are
equivalent; however, the XC can't evaluate (UNLESS (BOUNDP) SOMETHING)
with FOPCOMPILE, so don't try to: always use DEFPARAMETER to keep it simple.

A few places merited comments about this, as is could be considered remarkable,
but not enough so to remark about it every time. One noteworthy use occurs with
*QUEUED-PROCLAIMS* which mustn't get clobbered; its treatment in cold-init is
not an exception though; it's the rule: never overwrite in cold-init a variable
that was assigned in cold-load.