contrib/sb-cover: detect and emulate IN-PACKAGE when recording source maps
Standard code can be read independently of the current package.
However, our local package nickname extension (or some reader macros,
in theory) makes it necessary to be in the correct package for
SB-IMPL::READ-MAYBE-NOTHING to do the right thing.

Implement the same logic as SWANK: we scan each line of code as a
string, ook for lines that start with "(in-package " or
"(cl:in-package ", read that line, and find the package that
corresponds to the second element in the resulting list.

Tested with a new case in sb-cover/tests.lisp.