Remove bogus "#!+cmu (:pure nil)"
There might have been some point to having ":PURE NIL" in the past,
but it's clearly wrong, since obviously #!+cmu is and always has been false.

It would have been correct if a HAIRY-TYPE could reference mutable objects,
which it should not, nor can I see how it ever could. NEGATION-TYPE which came
from rev 58ff25d134 just borrowed the existing wrong conditional.

Incidentally, the two remaining uses of ":pure nil" to override the inherited
value of T from CTYPE seem ok or at least not wrong. MEMBER-TYPE is not
supposed to be mutable, but in the interest of safety, we might assume that
users could mutate the objects in a member-type.