Put undefined fun whitelist in package-data-list
This is a precursor to a change that will render it difficult if not impossible
to misuse "#+" or "#-" when "#!+" (resp "#!-") was intended and vice-versa.

Files that themselves are read in host mode (such as "compile-cold-sbcl") won't
have access to the target features as read suppression conditions. The features
will be present in SB-XC:*FEATURES* but without the char macros having been set.

Conversely, host features won't be testable with "#+/-" when building or running
the cross-compiler except by using the "(host-feature x)" escape syntax of the
target feature reader. In practice, such syntax is rarely called for any more.