Clean up a bunch of stuff around lp#540063 handling
Fixed 3 years ago, the solution for clearing autogenerated source-transforms
when functions get manually redefined left something to be desired - the DEFUN
macro had to look into the body to see whether it matched what "should be" the
sexpr for a DEFSTRUCT-generated thing.  A much simpler and less flaky approach
has DEFSTRUCT inform DEFUN which functions were generated. This patch achieves
that by adding a new macro SB-C:XDEFUN which takes one additional argument for
what kind of defstruct-generated thing a function is. It perhaps could have
been a new kind of declaration, but new declarations are less simple.

This also cleans up some supporting logic as follows:

* (INFO :FUNCTION :TYPE) can be only a CTYPE, or the symbol :GENERIC-FUNCTION,
  or sometimes during self-build, an s-expression instead of the CTYPE itself.
  The possibility of a DEFSTRUCT-DESCRIPTION is gone.

* We won't explicitly store an inline expansion for a defstruct ctor in a null
  environment when globally declaimed inline. It can always be reconstructed.

* Multiple non-BOA-constructors for the same defstruct all get transforms.
  (I don't know why you'd want this, but whatever.)

* PROCLAIMED-FTYPE is renamed GLOBAL-FTYPE since the type it returns
  has not necessarily been proclaimed.

* GLOBAL-FTYPE can return the type for any of DEFSTRUCT-generated thing,
  though they lack proclamations, which fixes one test case in sb-introspect
  as well as removes its special treatment of :COPIER functions.

Somewhat related, %DEFUN performs steps in the following order which is new:
   1. Potentially warn about redefinition
   2. %COMPILER-DEFUN
   3. (SETF FDEFINITION)
   4. Set inline expansion
It used to perform %COMPILER-DEFUN prior to the "maybe warn" which seems weird
because if the warning performed a nonlocal control transfer - like maybe
you disallow refinitions - then why was stuff only partway done?