Simplify APPLY-{mumble}-FIXUPS
Figure out whether or not to call FIND-LAYOUT based on whether we have
a layout instance or a symbol naming a layout in the FIXUP object.

I tried to make this 100% consistent by always creating the fixup using a
symbol, and always calling FIND-LAYOUT in apply fixups, but would you believe
that actually broke IR1?! Waaat? Well, the class system is complicated,
and a hand-wavy explanation is that any time you call FIND-LAYOUT, you're
actually causing effects upon the class system in a subtle way depending on
whether a classoid for the layout had existed or not.  This ended up harming
(among other things), DEFINE-CONDITION which creates a layout in a
:COMPILE-TOPLEVEL situation and then wants to find the created layout.
How could it possibly be *worse* to go from layout to name back to layout -
always treating layout fixups as though dumped / loaded - even when compiling
to memory? Well somehow it DOES harm things because the dumped layout was
the obsolete one, not a new one, which chokes right away. Srsly wtf.