macOS: Disable ud2-breakpoints, use 0xCC
The issue of not reliably firing signal handlers seems to fall under
the umbrella of problems described in 'mach-exception-handler-notes'.
While BSD signal handling may have had problems, it's hard to imagine that
mach exceptions did not work, considering that the comments alleging that
were written before the switch from BSD signals to catch_exception_raise.

Another point in favor of using 0xCC breakpoints is that MacOS 10.4.5
was released over 10 years ago, and a bug about failing to stop at breakpoints
would likely not have gone unaddressed for that length of time.

The final point is that it's nice to have both breakpoint instructions usable.