Try to fix more of the fun-end-breakpoint traps
All the -assem.S files were hardcoded to assume 6 boxed code header words
in the bpt-lra object. That included two words of zeros, so it should really
be reduced to 4 words; however, PPC seems to have more hidden assumptions
and 6 has to remain as 6. Mips can use 4 words and that works, though the
test suite is pretty crashy on the whole for reasons unrelated.
Sparc, alpha, and hppa are theoretically right but untestable by me.