Remove a duplicated defconstant
It's fine to parameterize the floating-point bit positions,
but it should be done in such a way as to segregate by the kind of
floating-point hardware the machine has, and not per backend
since - guess what - all our backends have the same parameters.

Eventually all of those constants should be pulled together
so that they needn't be repeated N times. This is just a start.

Also, rename define-float-dispatching-function which was misleading.
It's not merely another way of writing number-dispatch - it injects
logic pertinent to infinity/NaN testing.