Remove obsolete logic from genesis
When re-exporting symbols from packages other than their home, we used to
refrain from affecting certain symbols (LDB, BYTE, etc) that were
shadowing-imported into every SB-mumble package but not intended to be
replicated to all packages in the target. We no longer shadowing-import, but
instead simply omit some CL symbols when building the XC-STRICT-CL package.
As such is it neither required nor correct to ignore symbols that are homed
in SB-XC and also present and external in another package.