Speed up stream.impure and don't write into source tree
Use threads (if available) instead of run-program for one of the tests.
In our sandboxed environment, it went from ~270 seconds to 6.4 seconds
probably from a combination of fewer filesystem calls as well as scheduler
penalties on processes that fork. A normal run shows less dramatic
of an improvement: 8 seconds to 3 seconds, which is not nothing.

Also fix a horrible style issue - MOCK-FD-STREAM-N-BIN was both the name
of a slot accessor and a hand-written defun. Same for MOCK-FD-STREAM-IN.