Fix parallel-exec and improve purity of pure tests
Change PURE-RUNNER to randomize the package if and only if the filename
does not contain "impure". This makes it feasible again to use on all files
regardless of the pure/impure dichotomy.

In pure tests, disallow certain defining macros which instate persistent
environment changes which can't be undone by DELETE-PACKAGE.

Rename the backq tests to '.pure' since they are side-effect free.