Clear more caches in between steps of parallelized build
Specifically, this unscrews SB-XC:SUBTYPEP and some other things
after parallelized make-host-1.

The VALUES-SPECIFIER-TYPE exceptional definition used only in the x-compiler
never evicts entries ("The xc host uses an ordinary hash table for memoization."
originally from rev fb76d491) and caused invariant violations regarding
type-equivalence of interned ctypes.

Yielding an interned object from any ctype constructor returns the result of
a (LOAD-TIME-VALUE) form that should have produced an object not TYPE= to any
other object at the time of computation. Things went haywire if it did produce
a TYPE= object, because for interned ctypes there should be no other parse of
any given (up to EQUAL) type specifier. Thus, comparing a "new" parse to a
cached parse could say NIL instead of T in the quick failure case that assumes
that not-EQ implies not TYPE=. The fault was not in the constructor,
but rather lack of use of it due to inconsistent state.

The reason things weren't more broken is that prior to executing make-host-2
we do a total reload of artifacts from make-host-1. However, prior to this fix,
the newly added assertions would fail because they run before the total reload.