Allow reference of floating-point constants in cold load
To solve an ordering problem, use the host's proxy of the target value (NOT the
host's value) if the target value has not been assigned into the cold core yet.
There shouldn't technically be anything referenced out-of-order with respect
to its definition, except that certain defining forms are deferred to cold-init,
while a using form occurs in cold-load, hence the problem.