Relax specifier-type transform a little
The only type spec that can't be eagerly translated to a type object at
cross-compile time is CLASS (which means that the only type specifier that
doesn't work is TYPE-SPECIFIER).  The reason it worked without CLASS in the
blacklist before is that TYPE-SPECIFIER also includes CLASSOID which was
enough to inhibit the transform.