Combine layout-length, depthoid, flags into one slot if #+64-bit
This is in preparation for a change which needs to add new slots in LAYOUT.
We don't generally care too much about small increases in structure length,
however, experience has shown that adding physical slots to LAYOUT is a prime
cause of performance degradation. There is no additional overhead to reading
the depthoid value, as it is a naturally-aligned 4-byte quantity.