Set vector weakness bit before filling in contents
The allocator vop and/or the GC may need to be advised of favoritism
for certain pages holding only weak vectors. That doesn't work if you
manipulate the weakness bit after creating a plain simple vector.
We don't actually exploit this yet in any way.

And the code was arguably imperfect by populating the vector prior to flagging
as weak. It can't have made any difference, because for practical purposes the
constructor function holds strong references to the initializer value.

A visual check of all the backends to see whether they were happy
storing a 9-bit vector widetag suggests that none were bothered by it
except for x86, which now is fine with that.