Adhere to C language spec
Program counters need to be either uword_t or char* and not lispobj* because
the C compiler may assume that lispobj* is properly aligned for its type.
But x86 machines have no instrution alignment, and for 64-bit risc machines
that use 4 byte instruction alignment we were in violation as well.