Fix bug in pscav_later(), and use it as intended
It only worked if count was strictly less than LATERMAXCOUNT or an exact
multiple thereof. That was due to the "if (count > LATERMAXCOUNT)"
precluding use of the 'else' case for any remainder in the 'then' branch.

However, the whacky queue representation was a premature optimization
that we can dispense with. Given that simple-funs now have no boxed words,
it takes only 2 integers to store the deferred scavenge range per code object
whereas it took (1 + 2 * N_simple_funs) integers in the former approach,
and there is no sense at all in trying to special-case a range of 1.