Add metadata indicating data files per test file
This is important for sandboxed execution where each executor receives exactly
the files that it should and no more. To be slightly compatible in a normal
local testing scenario, the test runner will attempt to generate the same error
as would be signaled by OPEN in the sandbox for undeclared inputs. Listing
of '.so' and '.fasl' in the manifest doesn't do anything locally, but it does
in a real sandbox where binary artifacts can also be inaccessible.