Fix some really wonky math in hash-table sizing
This code seems to have frequently confused the following:
 - the number of k/v pairs that could be stored
 - the number of buckets
 - the sizes of the vectors after adding the fixed overhead

So for example the min-size of 17 actually meant 16 k/v pairs which is a
rather lame interpretation for that constant. And in HASH-TABLE-NEW-VECTORS
we computed the new number of k/v cells from the old number multiplied by 1.5,
and rounded to a power of 2, which basically meant doubling; and then we used
the same number as the bucket count, but because of the physical<->logical cell
count discrepancy, we end up with one fewer logical cell, so consequently we
always had (2^n-1) items distributed across 2^n bins and so our tables were
allowed to operate at a load factor of 1 which was not really the plan.