Improve PUTHASH and MAPHASH
Create specializations of PUTHASH for each of the 4 standard tests
and one for any user-defined test.

In MAPHASH and WITH-HASH-TABLE-ITERATOR, don't reload the data vector on each
iteration. It turns out that was to support a certain implementation detail
of PUTHASH where it might assign a new table unnecessarily (assuming the
situation permitted PUTHASH during iteration, i.e. to alter the value for the
current key). If the table was 100% full, then PUTHASH installed a new
kv-vector right away even though the key would have been found.

And as luck would have it, several 'mop.impure' tests randomly provoked a
problem when I tried removing the reload of kv-vector without the entirety
of this change.  Indeed they created a table whose kv-vector was full,
and MAPHASHed over it, and used (SETF GETHASH) during the maphash, and crashed
when they observed the 0s that REHASH scribbles over the old vector.
So this change avoids resizing in PUTHASH unless there really isn't room.