Remove unwarranted and inexplicable WITHOUT-GCING
If I had to guess, that change concerned the global malloc lock.
It's likely that dladdr uses malloc, and there are ways to induce GC to
malloc, but only via fprintf now when logging is enabled. There might have
been other ways before. The permanent solution is to avoid fprintf.