Enable stably hashed EQ hash tables
To get this behavior, you can specify :TEST 'EQ :HASH-FUNCTION F
where F produces stable hashes.  F is not permitted to return any
address-based hashes though - it's strictly one or the other.

Change the GF tables in method-combination objects to use stable hashes.
And use EQ, not EQL because it is almost always wrong to use EQL
as the comparator with weak-keyed tables.