Remove the length byte from fdefn headers
and for immobile fdefns store the generation byte where the length was
which gives some room to encode a longer JMP instruction, which in turn
helps eliminates the restriction of using a rel32 operand.

For consistency, all the architectures use a constant 4 for FDEFN_SIZE
and all avoid explicitly representing it in the header.