Implement string output streams of base-char
As suggested by Nikodemus in rev 12b7e571c9a7.

Also, for string output streams that accept any character but prefer to
return a base-string, use a better test of whether the result contains
non-base characters: examine strings when written instead of pre-scanning
the buffers in GET-OUTPUT-STREAM-STRING.  If only base-strings were written,
an extra traversal to test for Unicode characters is avoided.