immobile-code: Store absolute ptr in fdefn-raw-addr
Encoding a JMP instruction was too limiting. With this change we can place
simple-funs anywhere in memory without needing a trampoline to ensure proximity
to the corresponding fdefn. Space defragmentation is easier as well.

Closures and FINs which require a simplifying trampoline will allocate
the trampoline in dynamic space so that we no longer create a huge number
of 6-word objects in immobile code space.

This change should also allow placing funcallable-instances with
self-contained trampolines in dynamic space, but does not do so yet.

As far as call convention goes, there is still a limitation that callers
in immobile space must be near enough an fdefn to use a 'call rel32'
instruction format.