Simplify the storage of "fancy" fdefn names
Implement a custom hashset for fdefns that hold PCL method names, do not use
globaldb's *info-environment* table.
If nothing else, this consumes less space than a hashtable, because each fdefn
stores its own name, making an external mapping of name to fdefn redundant.

The rationale for prohibiting the storage of anything in globaldb is that it
contributes to persistence of garbage even after DELETE-PACKAGE and other
operations that should scrub data.  I have not yet figured out how to
actually achieve the intent of being able to drop all methods associated
with symbol FOO when FOO becomes otherwise unreachable, but the comments
sketch out a couple ideas.

It seems like *INFO-ENVIRONMENT* could be entirely deleted at this point.