allow redefinitions of erroneous classes
If a class definition fails because of an arity mismatch in a reader
or writer method, subsequent attempts to redefine the class correctly
will fail through attempts to remove the (non-existent) method -- the
errorp argument to FIND-METHOD (and our internal function GET-METHOD
which implements it) does not control arity mismatch errors, only
nonexistence errors.  Check for the generic function having the
correct arity before attempting to remove reader/writer/boundp
methods.  Thanks to 3b on #sbcl for the report.