Move OPEN restarts into separate function; add :IF-EXISTS :APPEND
Most of the things requested in lp#806398 were already in place. This
change only adds the APPEND restart for existing files which retries
as if :IF-EXISTS :APPEND had been supplied.

The other part of this change is reducing the amount of code in OPEN
by

* Using LOOP/RETURN instead of BLOCK+TAGBODY/GO

* Performing error signaling and restart creation in a separate
  function %OPEN-ERROR

The new restart is tested in the (open :if-exists restart) test.

fixes lp#806398