update to Unicode 8.0.0
(except for Confusables, which got updated to 12.0.0 at some point)

Unicode 8.0.0 is not very recent (though more recent than Unicode
7.0.0, which we supported previously).  A large point of this exercise
was to discover the irritating manual fixups to SBCL source and to the
Unicode data files which following commits will lessen the need for,
after which an update to Unicode 9, 10, 11, or 12 should be more
straightforward.

In the process, fix a bug in line breaking after a hebrew letter
followed by a hyphen.  (This was a bug in our Unicode 7.0.0 support,
but the test case was only included with Unicode 8.0.0)