Remove horrible semantics
block_signals() and unblock_signals() each had two completely different
behaviors - operating on a signal mask in memory, or instructing the OS
to change the mask. Combined with the opaque parameter names of 'what' and
'where', this was not a very mnemonic API. So fix it by splitting those
functions into pieces that do what their names imply.