translate a not-quite-ASCII data file to ASCII
This is not an elegant remedy to the fact that we can assume almost
nothing about our host compiler's character encoding.  For maximum
portability, anything that is read as text in the course of the SBCL
build should constrain itself to the standard-char repertoire, which
most but not all Unicode data files do: unfortunately, CaseFolding has
some explanatory comments at its head, including a brief discussion of
the fact that a single character, such as the German eszet, can upcase
to more: which would be fine except that the data file contains a
UTF-8 encoded eszet.

Treat the CaseFolding.txt data file as a binary artifact, and rewrite
it such that it can be read as text in the build.