Avoid false positives from C code compiled with -fsanitize=memory
We don't update stack shadow words from lisp, so depending on whether C calls
occurred (which do poison/unpoison parts of the stack), it could be impossible
to backtrace in ldb because the sanitizer would report that x6_call_context
was reading uninitialized memory. Same goes for the 'dump' command.