Add 'search' command to ldb
Revision e1676d5093b1 deleted a similarly named command, but that 'search'
was a fragile implementation of a useless notion of searching that scanned
a specified address range looking for a given widetag.
This 'search' command is helpful when you don't feel like firing up a debugger
to call search_all_gc_spaces() to decipher an interior pointer.