build CaseFolding data identically on all hosts
Tighten the conversion of the Unicode-supplied data file to assert
that we convert exactly one high-octet pair, an eszet, and that
there are no other high-octet characters in the file.  Fail with
errors if these are not true: if they aren't, then Unicode has
probably put other important information in there that the intrepid
SBCL developer should read.  (Or the file has been encoded using
EBCDIC-high or something.)

The sandboxed build issue referred to in a previous commit message
is solved another way.