Add yet another encapsulation mechanism
Allow a simple-fun 'self' to point to a function outside the containing code.
In particular, the external function can implement a tracing wrapper which
preserves object identity of the traced function and traces through unnamed
calls. This is pretty much what we already do for traced generic functions,
and removes a major use-case for the more complicated and un-threadsafe
breakpoint-based tracing. (It's possible to implement breakpoints threadsafely,
but that's a very tricky endeavor)

The TRACE macro does not yet use this facility. I suspect that doing so will
crash on the non-x86 platforms. The smoke test passes, but the real test is
whether an fdefn-fun and fdefn-raw-addr slot are allowed to be inconsistent.
They probably are not, because that will wreak havoc upon the LIP register
as used for named call.