x86-64: Further improve generated code for CASE on symbols
The object given to SYMBOL-HASH doesn't need to be a symbol - it just needs
opaque bits that might map back to itself in the key vector, which only happens
if it was in fact a symbol. So look for any pointer lowtag, read 1 word beyond
the untagged pointer (which for symbols is the hash) and mask that value.
When the argument type is derived as a symbol, then avoid untagging.