Use CASE in convert-more-entry
This exposed the sometimes utter horribleness of the hash generator.
Example: %MAKE-FD-STREAM takes 37 keywords. Rounded to a power of 2 that's 64.
The table can have double the minimum size which is 128. Collisions can add
up to one extra "stripe" of symbols, the first stripe containing every symbol
that didn't collide with another, and the second stripe holding all leftovers.
So 256 vector elements were needed to accept 37 symbols, totaling 86% waste.
Now it's down to one symbol vector of 128 elements, for only 63% waste,
plus the jump table which also has that much waste.

If we want to get insanely clever, a better hash generator could map N symbols
into 0..N-1 to shrink the jump table. There's an easy solution: use the index
in the stripe to pick out a small integer from an ancillary vector,
and use the small integers as CASE keys.
It's probably not worth adding another level of indirection though.