Fix GC bug re. jump tables
We must not look for a jump table until code->boxed_size slot holds a nonzero
value, or else we'd miscalculate where the unboxed words should be.
A GC triggered within allocate-code-object exposed a window of vulnerability,
though the x86 architectures obscure the bug due to the implicit pin of objects
in registers.  Other architectures do succumb to the bug, as I discovered
while implementing jump tables for ppc.  It's a bug even if hidden though.