Do some more TYPECASE to CASE conversion
If the multiway branch vop works and TYPECASE is over frozen structure
types, then map layout-clos-hash to a small integer and dispatch on that.
With some complications, this could be made to work for non-frozen types.

Also fix a general problem with multiway branch not figuring out when
a CASE is exhaustive with respect to the derived type of a numeric key.