Fix natural loop computation.
The loop computation was incorrect, since the smallest set of blocks
comprising a loop cannot be attained by a forward walk from the
header, as the tail must be reachable from all loop blocks in the
graph with the header block removed. Instead, a backward walk from the
tails is needed to find the minimal set of blocks constituting a
natural loop, without introducing blocks which are not on the paths
from the header block to tail blocks.