Compute deterministic layout hashes
They don't have to be pseudorandom. Named classes can use a name-based hash.
In fact the hash can stay the same when a class gets redefined and the old
layout's hash is clobbered to 0. PCL cache lookup has to compare layouts by
EQ eventually, because it must allow for hash collisions in any case.