Fix warning with recent CLISP
Without this change, it prints:

  ; file: /tmp/sbcl/src/compiler/ir2opt.lisp
  ; in: ; DEFOPTIMIZER ; (VOP-OPTIMIZE BRANCH-IF)
  ;     (SB-C:DEFOPTIMIZER (SB-C::VOP-OPTIMIZE SB-C:BRANCH-IF) (SB-C:BRANCH-IF)
  ;    (LET ((SB-C::PREV #)) (WHEN (AND SB-C::PREV #) (LET # #))))
  ; caught WARNING:
  ;   ; Reference to WHAT is implementation-dependent, per ANSI CL 6.1.1.4.
  ** - Continuable Error
  FAILURE-P was set when creating "obj/from-xc/src/compiler/ir2opt.lisp-obj".

It seems like it's doing a sloppy job at semantic analysis. It thinks that the use
of WHAT inside the initializing form of the FOR .. IN references the same WHAT
that is bound as the FOR variable, or something similarly nonsensical.

It's easy enough to make the warning go away with a trivial rename.