Avoid some work in zero_pages()
If the memory would be set with memset() and then almost immediately subject
to memcpy(), then skip the memset.

When running the regression suite, deliberately write garbage to any page that
was eligibile to skip zero-initialization exactly when that would have occurred,
as a check that both the GC and ALLOCATE-CODE-OBJECT are insensitive to whether
a page held old garbage.

The remarks above zero_pages() sketch out some further ideas.
(This change is more comments than code)