Rearrange disassembler
This change moves things around so target-only code is not compiled to
execute the host.  No functional change.

Granted, that are folks who would like to see, in addition to pre-assembly
code in the trace file during cross-compilation, a full disassembly of the
assembler's output buffer. I don't know of an easy way to do that.