Require :sb-dynamic-core for arm64
I assumed that this would help find the cause of crashing with #-sb-thread,
and miraculously the change actually seems to have entirely fixed the problem.
Given that observation, I surmised that the bug lay with misuse of LR as a
temp register which this change removes. Further confirmation was obtained by
reverting the patch and merely adding (LOAD-IMMEDIATE-WORD ,LIP 0) after each
inline allocation. Simply clearing LIP made the random crashing disappear.

I don't know if there was a point in the past when the non-thread build was
perfectly reliable despite the allocation macro being relatively unchanged.
It didn't seem sufficiently interesting a question for me to answer, but in
the interest of code health, I think the LIP argument should be removed.