Cause make-config to run successfully by default
i.e. do not fail if the host lisp isn't found, barring any other failures.

Intuitively the check for a host lisp belongs in 'make.sh', but the rationale
for putting it in make-config is that failing sooner makes sense if the build
can't begin. But it was an incorrect behavior for make-config when run at the
shell prompt while trying to diagnose problems in that script.