Remove one of many netbsd hacks
No extra guard of (and (not netbsd)) is needed, now that
:OS-PROVIDES-GETPROTOBY-R is definitely absent from *FEATURES*.

As to why someone removed the locking around getprotobyname, that looked
wrong to me so I put the mutex back in, as it should be for every platform
that lacks the reentrant library call. If NetBSD had made getprotobyname
(the non-reentrant one) return data into a thread-local buffer, that would
explain things. However, the manual page contains the usual caveat:
 "These functions use a static data space; if the data is needed for
  future use, it should be copied before any subsequent calls overwrite it."
https://netbsd.gw.com/cgi-bin/man-cgi?getprotobyname+3+NetBSD-current